Refactoring
Fix structure.
