+++ internal/listener/publisher_mock_test.go
@@ -2,10 +2,9 @@ package listener
 
 import (
 	"context"
+	"github.com/ihippik/wal-listener/v2/internal/publisher"
 
 	"github.com/stretchr/testify/mock"
-
-	"github.com/ihippik/wal-listener/v2/publisher"
 )
 
 type publisherMock struct {
