+++ internal/listener/transaction/parser_test.go
@@ -1,8 +1,9 @@
-package listener
+package transaction
 
 import (
 	"bytes"
 	"encoding/binary"
+	"github.com/ihippik/wal-listener/v2/listener"
 	"io"
 	"log/slog"
 	"reflect"
@@ -461,17 +462,17 @@ func TestBinaryParser_getBeginMsg(t *testing.T) {
 func TestBinaryParser_ParseWalMessage(t *testing.T) {
 	type args struct {
 		msg []byte
-		tx  *WalTransaction
+		tx  *WAL
 	}
 
 	logger := slog.New(slog.NewJSONHandler(io.Discard, nil))
-	metrics := new(monitorMock)
+	metrics := new(listener.monitorMock)
 
 	tests := []struct {
 		name    string
 		args    args
 		wantErr bool
-		want    *WalTransaction
+		want    *WAL
 	}{
 		{
 			name:    "empty data",
@@ -487,9 +488,9 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					0, 0, 0, 0, 0, 0, 0, 0,
 					0, 0, 0, 5,
 				},
-				tx: NewWalTransaction(logger, nil, metrics),
+				tx: NewWAL(logger, nil, metrics),
 			},
-			want: &WalTransaction{
+			want: &WAL{
 				pool:          nil,
 				log:           logger,
 				LSN:           7,
@@ -510,7 +511,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					0, 0, 0, 0, 0, 0, 0, 8,
 					0, 0, 0, 0, 0, 0, 0, 0,
 				},
-				tx: &WalTransaction{
+				tx: &WAL{
 					log:           logger,
 					LSN:           7,
 					monitor:       metrics,
@@ -518,7 +519,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					RelationStore: make(map[int32]RelationData),
 				},
 			},
-			want: &WalTransaction{
+			want: &WAL{
 				log:           logger,
 				LSN:           7,
 				monitor:       metrics,
@@ -553,7 +554,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					0, 0, 0, 23,
 					0, 0, 0, 1,
 				},
-				tx: &WalTransaction{
+				tx: &WAL{
 					log:           logger,
 					LSN:           3,
 					monitor:       metrics,
@@ -562,7 +563,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					RelationStore: make(map[int32]RelationData),
 				},
 			},
-			want: &WalTransaction{
+			want: &WAL{
 				log:        logger,
 				monitor:    metrics,
 				LSN:        3,
@@ -608,7 +609,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					0, 0, 0, 6,
 					49, 48,
 				},
-				tx: &WalTransaction{
+				tx: &WAL{
 					monitor:    metrics,
 					log:        logger,
 					LSN:        4,
@@ -631,7 +632,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					},
 				},
 			},
-			want: &WalTransaction{
+			want: &WAL{
 				monitor:    metrics,
 				log:        logger,
 				LSN:        4,
@@ -701,7 +702,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					0, 0, 0, 2,
 					56, 48,
 				},
-				tx: &WalTransaction{
+				tx: &WAL{
 					log:        logger,
 					monitor:    metrics,
 					LSN:        4,
@@ -724,7 +725,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					},
 				},
 			},
-			want: &WalTransaction{
+			want: &WAL{
 				monitor:    metrics,
 				log:        logger,
 				LSN:        4,
@@ -792,7 +793,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					0, 0, 0, 2,
 					55, 55,
 				},
-				tx: &WalTransaction{
+				tx: &WAL{
 					monitor:    metrics,
 					log:        logger,
 					LSN:        4,
@@ -815,7 +816,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					},
 				},
 			},
-			want: &WalTransaction{
+			want: &WAL{
 				monitor:    metrics,
 				log:        logger,
 				LSN:        4,
@@ -868,7 +869,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					0, 0, 0, 2,
 					55, 55,
 				},
-				tx: &WalTransaction{
+				tx: &WAL{
 					monitor:    metrics,
 					log:        logger,
 					LSN:        4,
@@ -891,7 +892,7 @@ func TestBinaryParser_ParseWalMessage(t *testing.T) {
 					},
 				},
 			},
-			want: &WalTransaction{
+			want: &WAL{
 				monitor:    metrics,
 				log:        logger,
 				LSN:        4,
