+++ internal/listener/transaction/pg_type.go
@@ -1,4 +1,4 @@
-package listener
+package transaction
 
 // PostgreSQL OIDs
 // https://github.com/postgres/postgres/blob/master/src/include/catalog/pg_type.dat
