+++ internal/listener/transaction/protocol.go
@@ -1,4 +1,4 @@
-package listener
+package transaction
 
 import (
 	"time"
