+++ internal/listener/transaction/wal.go
@@ -1,37 +1,25 @@
-package listener
+package transaction
 
 import (
 	"context"
+	"errors"
+	"github.com/ihippik/wal-listener/v2/internal/publisher"
 	"log/slog"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/goccy/go-json"
 	"github.com/google/uuid"
-
-	"github.com/ihippik/wal-listener/v2/publisher"
 )
 
-// ActionKind kind of action on WAL message.
-type ActionKind string
-
-// kind of WAL message.
-const (
-	ActionKindInsert ActionKind = "INSERT"
-	ActionKindUpdate ActionKind = "UPDATE"
-	ActionKindDelete ActionKind = "DELETE"
-)
-
-type transactionMonitor interface {
+type monitor interface {
 	IncFilterSkippedEvents(table string)
 }
 
-// WalTransaction transaction specified WAL message.
-type WalTransaction struct {
+// WAL transaction specified WAL message.
+type WAL struct {
 	log           *slog.Logger
-	monitor       transactionMonitor
+	monitor       monitor
 	LSN           int64
 	BeginTime     *time.Time
 	CommitTime    *time.Time
@@ -40,11 +28,13 @@ type WalTransaction struct {
 	pool          *sync.Pool
 }
 
-// NewWalTransaction create and initialize new WAL transaction.
-func NewWalTransaction(log *slog.Logger, pool *sync.Pool, monitor transactionMonitor) *WalTransaction {
+var errRelationNotFound = errors.New("relation not found")
+
+// NewWAL create and initialize new WAL transaction.
+func NewWAL(log *slog.Logger, pool *sync.Pool, monitor monitor) *WAL {
 	const aproxData = 300
 
-	return &WalTransaction{
+	return &WAL{
 		pool:          pool,
 		log:           log,
 		monitor:       monitor,
@@ -53,114 +43,23 @@ func NewWalTransaction(log *slog.Logger, pool *sync.Pool, monitor transactionMon
 	}
 }
 
-func (k ActionKind) string() string {
-	return string(k)
-}
-
-// RelationData kind of WAL message data.
-type RelationData struct {
-	Schema  string
-	Table   string
-	Columns []Column
-}
-
-// ActionData kind of WAL message data.
-type ActionData struct {
-	Schema     string
-	Table      string
-	Kind       ActionKind
-	OldColumns []Column
-	NewColumns []Column
-}
-
-// Column of the table with which changes occur.
-type Column struct {
-	log       *slog.Logger
-	name      string
-	value     any
-	valueType int
-	isKey     bool
-}
-
-// AssertValue converts bytes to a specific type depending
-// on the type of this data in the database table.
-func (c *Column) AssertValue(src []byte) {
-	var (
-		val any
-		err error
-	)
-
-	if src == nil {
-		c.value = nil
-		return
-	}
-
-	strSrc := string(src)
-
-	const (
-		timestampLayout       = "2006-01-02 15:04:05"
-		timestampWithTZLayout = "2006-01-02 15:04:05.999999999-07"
-	)
-
-	switch c.valueType {
-	case BoolOID:
-		val, err = strconv.ParseBool(strSrc)
-	case Int2OID, Int4OID:
-		val, err = strconv.Atoi(strSrc)
-	case Int8OID:
-		val, err = strconv.ParseInt(strSrc, 10, 64)
-	case TextOID, VarcharOID:
-		val = strSrc
-	case TimestampOID:
-		val, err = time.Parse(timestampLayout, strSrc)
-	case TimestamptzOID:
-		val, err = time.ParseInLocation(timestampWithTZLayout, strSrc, time.UTC)
-	case DateOID, TimeOID:
-		val = strSrc
-	case UUIDOID:
-		val, err = uuid.Parse(strSrc)
-	case JSONBOID:
-		var m any
-
-		if src[0] == '[' {
-			m = make([]any, 0)
-		} else {
-			m = make(map[string]any)
-		}
-
-		err = json.Unmarshal(src, &m)
-		val = m
-	default:
-		c.log.Debug(
-			"unknown oid type",
-			slog.Int("pg_type", c.valueType),
-			slog.String("column_name", c.name),
-		)
-
-		val = strSrc
-	}
-
-	if err != nil {
-		c.log.Error(
-			"column data parse error",
-			slog.String("err", err.Error()),
-			slog.Int("pg_type", c.valueType),
-			slog.String("column_name", c.name),
-		)
-	}
-
-	c.value = val
-}
-
 // Clear transaction data.
-func (w *WalTransaction) Clear() {
+func (w *WAL) Clear() {
 	w.CommitTime = nil
 	w.BeginTime = nil
 	w.Actions = nil
 }
 
-// CreateActionData create action  from WAL message data.
-func (w *WalTransaction) CreateActionData(
+func (w *WAL) RetrieveEvent(event *publisher.Event) {
+	w.pool.Put(event)
+}
+
+func (w *WAL) getPoolEvent() *publisher.Event {
+	return w.pool.Get().(*publisher.Event)
+}
+
+// CreateActionData create action from WAL message data.
+func (w *WAL) CreateActionData(
 	relationID int32,
 	oldRows []TupleData,
 	newRows []TupleData,
@@ -213,7 +112,7 @@ func (w *WalTransaction) CreateActionData(
 
 // CreateEventsWithFilter filter WAL message by table,
 // action and create events for each value.
-func (w *WalTransaction) CreateEventsWithFilter(ctx context.Context, tableMap map[string][]string) <-chan *publisher.Event {
+func (w *WAL) CreateEventsWithFilter(ctx context.Context, tableMap map[string][]string) <-chan *publisher.Event {
 	output := make(chan *publisher.Event)
 
 	go func(ctx context.Context) {
@@ -235,7 +134,8 @@ func (w *WalTransaction) CreateEventsWithFilter(ctx context.Context, tableMap ma
 				data[val.name] = val.value
 			}
 
-			event := w.pool.Get().(*publisher.Event)
+			event := w.getPoolEvent()
+
 			event.ID = uuid.New()
 			event.Schema = item.Schema
 			event.Table = item.Table
