+++ internal/listener/transaction/wal_test.go
@@ -1,4 +1,4 @@
-package listener
+package transaction
 
 import (
 	"io"
@@ -135,7 +135,7 @@ func TestWalTransaction_CreateActionData(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			w := WalTransaction{
+			w := WAL{
 				log:           logger,
 				LSN:           tt.fields.LSN,
 				BeginTime:     tt.fields.BeginTime,
