+++ internal/publisher/event.go
@@ -2,10 +2,10 @@ package publisher
 
 import (
 	"fmt"
+	"github.com/ihippik/wal-listener/v2/internal/config"
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/ihippik/wal-listener/v2/config"
 )
 
 // Event structure for publishing to the NATS server.
