+++ internal/publisher/rabbit.go
@@ -3,9 +3,9 @@ package publisher
 import (
 	"context"
 	"fmt"
+	"github.com/ihippik/wal-listener/v2/internal/config"
 
 	"github.com/goccy/go-json"
-	"github.com/ihippik/wal-listener/v2/config"
 	"github.com/wagslane/go-rabbitmq"
 )
