--- listener/errors.go
@@ -1,14 +1,0 @@
-package listener
-
-import "errors"
-
-// Variable with connection errors.
-var (
-	errReplConnectionIsLost = errors.New("replication connection to postgres is lost")
-	errConnectionIsLost     = errors.New("db connection to postgres is lost")
-	errMessageLost          = errors.New("messages are lost")
-	errEmptyWALMessage      = errors.New("empty WAL message")
-	errUnknownMessageType   = errors.New("unknown message type")
-	errRelationNotFound     = errors.New("relation not found")
-	errReplDidNotStart      = errors.New("replication did not start")
-)
